+++ brain_app/src/astar-jps/IndexPriorityQueue.h
@@ -2,7 +2,7 @@
 #define PRIORITYQUEUE_H_
 
 typedef struct item {
-	double priority;
+	float priority;
 	int value;
 } item;
 
@@ -14,10 +14,10 @@ typedef struct queue {
 	unsigned int indexAllocated;
 } queue;
 
-void insert (queue *q, int value, double priority);
+void insert (queue *q, int value, float priority);
 void deleteMin (queue *q);
 item *findMin (const queue *q);
-void changePriority (queue *q, int ind, double newPriority);
+void changePriority (queue *q, int ind, float newPriority);
 void delete (queue *q, int ind);
 int priorityOf (const queue *q, int ind);
 int exists (const queue *q, int ind);
